fix: preserve decimal timestamp

+++ backVLA/springboot/demo/src/main/java/com/example/demo/api/dto/SensorsDTO.java
@@ -15,6 +15,8 @@ public class SensorsDTO {
     private GPSDTO gps;
     private Integer esp_now_channel;
     private String mac_address;
-    private Long timestamp;
+    // `timestamp` do simulador vem como valor decimal (segundos com fração).
+    // Usar Double para não perder a parte fracionária ao desserializar.
+    private Double timestamp;
     private LancamentoEntity lancamento;
 }
